+++ src/main.rs
@@ -1,6 +1,7 @@
 use bevy::prelude::*;
 mod components;
 mod gui;
+mod scenario;
 mod physics;
 
 use components::{body, camera};
@@ -11,7 +12,8 @@ pub const DT: f32 = 1.0 / 60.0;
 // This affects the size of every elements in the physics engine, by multiplying all the length-related quantities by
 // the physics_scale factor. This should likely always be 1.0 in 3D. In 2D, this is useful to specify a
 // “pixels-per-meter” conversion ratio.
-pub const SCALE: f32 = 1.0;
+pub static SCALED_LENGTH: f32 = 1.0e-6; // km
+pub static SCALED_MASS: f32 = 1.0e-6;  // kg
 
 fn main() {
     App::new()
@@ -19,7 +21,7 @@ fn main() {
             DefaultPlugins.set(WindowPlugin {
                 primary_window: Some(Window {
                     #[cfg(not(target_arch = "wasm32"))]
-                    resolution: bevy::window::WindowResolution::new(1920.0, 1080.0),
+                    resolution: bevy::window::WindowResolution::new(600.0, 400.0),
                     fit_canvas_to_parent: true,
                     prevent_default_event_handling: false,
                     canvas: Some("#app".to_owned()),
@@ -28,26 +30,21 @@ fn main() {
                 ..default()
             }),
             // Interface
-            gui::BaseUiPlugin,
-            gui::hud::HudPlugin,
-            gui::orbit_prediction::OrbitPredictionPlugin,
-            gui::select::SelectionPlugin,
-            gui::debug::DebugUiPlugin, // requires BaseUiPlugin
-            // Camera
+            gui::GuiPlugin,
             camera::CameraPlugin,
-            //Physics
-            physics::PhysicsPlugin,
+            // Physics
+            physics::PhysicsSchedulePlugin,
             physics::GravityPlugin,
+            // Scene
+            scenario::ScenarioPlugin,
         ))
         // Resources
         .insert_resource(ClearColor(Color::BLACK))
         .insert_resource(AmbientLight {
-            color: Color::NONE,
-            brightness: 0.0,
+            color: Color::WHITE,
+            brightness: 0.5,
         })
-        .insert_resource(physics::PhysicsSettings::delta_time(1.0 / 60.0))
-        // Spawn bodies
-        .add_systems(Startup, body::setup_scene)
+        .insert_resource(physics::PhysicsSettings::delta_time(DT))
         .add_systems(First, body::add_materials)
         .run();
 }
